pyproject.toml
[tool.poetry]
name = "httpx-iap"
version = "0.1.0"
description = "A Google IAP extension for httpx"
authors = ["Evan Pikulski <epikulski@gmail.com>"]
license = "MIT"
packages = [
    { include = "httpx_iap", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
httpx = "^0.18.2"
PyJWT = "^2.1.0"
cryptography = "^3.4.7"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
pylint = "^2.9.5"
mypy = "^0.910"
black = "^21.7b0"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
pytest-httpx = "^0.12.0"
pytest-cov = "^2.12.1"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
